PtsTo class: now works
Just one type annotation needed in library, and it makes sense. Adding
instances for ref, gref, box, vec, array, higher array.

It's definitely not ideal to have to mark everything pulse_unfold, but
that seems easy to fix: we need to tag instances and methods with a
particular attribute and add it to the normalization call done by Pulse.