rewrite each: make sure to normalize/unfold each side
This fixes the attached test case, where we try to rewrite every `bar`
in the context into `baz`, but since `bar` is unfold, it will never
appear as such, making the rewrite do nothing.

This happens pretty often for operators like `SizeT.(+)`, which is just
defined as `SizeT.add`, with an unfold qualifier. Without this patch,
one cannot use these operators in the LHS of a rewrite and must use the
underlying names.