Pulse.Checker.Abs: tweak checking of expected type of abstractions
The current check is going via norm_st_typing which ends up doing a
syntactic check over a normalized version of the computed type and
annotation. This is needlessly brittle, and breaks in everparse in some
fancy function types. This can just be replaced by a call to
t_check_equiv.

I couldn't easily write a test for this though... Even in everparse,
the two terms that fail to be syntactically equal appear exactly the
same when printed. Using --ugly, I see differently-nested arrows, but
going via the reflection API should make that not be a problem.